divided by components
