+++ package.json
@@ -13,13 +13,10 @@
     },
     "dependencies": {
         "@types/body-parser": "^1.19.2",
-        "@types/http-errors": "^1.8.2",
         "axios": "^0.27.2",
         "axios-mock-adapter": "^1.21.2",
-        "coinmarketcap-api": "^3.1.1",
         "dotenv": "^16.0.2",
         "express": "^4.18.1",
-        "http-errors": "^2.0.0",
         "node-fetch": "^3.2.10",
         "node-ts-cache": "^4.4.0",
         "node-ts-cache-storage-memory": "^4.4.0",
