+++ src/app.ts
@@ -2,8 +2,7 @@ import express from 'express';
 import bodyParser from 'body-parser';
 
 import config from './config';
-import router from './routes/router';
-import { errorHandler } from './middlewares/error.middleware';
+import router from './router';
 
 export default class App {
     public app: express.Application;
