+++ src/components/rate/__tests__/unit/binance.test.ts
@@ -1,19 +1,17 @@
-import CoinbaseRateProvider from '../../../services/rate/providers/coinbase';
-import config from '../../../config';
+import BinanceRateProvider from '../../services/providers/binance';
+import config from '../../../../config';
 
 import axios from 'axios';
 import MockAdapter from 'axios-mock-adapter';
 
-describe('coinbase test', () => {
+describe('binance test', () => {
     test('provider should return rate', async () => {
         const mock = new MockAdapter(axios);
-        mock.onGet(config.COINBASE_ENDPOINT).reply(200, {
-            data: {
-                amount: 1000,
-            },
+        mock.onGet(config.BINANCE_ENDPOINT).reply(200, {
+            price: 1000,
         });
 
-        const rateClient = new CoinbaseRateProvider();
+        const rateClient = new BinanceRateProvider();
         const rate = await rateClient.getRate();
 
         expect(typeof rate).toBe('number');
@@ -24,7 +22,7 @@ describe('coinbase test', () => {
         const mock = new MockAdapter(axios);
         mock.onGet('errorURL').reply(400);
 
-        const rateClient = new CoinbaseRateProvider();
+        const rateClient = new BinanceRateProvider();
         const expectedError = async () => await rateClient.getRate();
 
         return expect(expectedError()).rejects.toThrowError();
