+++ src/components/rate/services/providers/binance.ts
@@ -1,15 +1,15 @@
 import axios from 'axios';
-import config from '../../../config';
+import config from '../../../../config';
 
 import AbstractRateProvider, { IRateProvider } from './provider.interface';
 
-export default class CoinbaseRateProvider extends AbstractRateProvider implements IRateProvider {
-    private readonly url: string = config.COINBASE_ENDPOINT;
+export default class BinanceRateProvider extends AbstractRateProvider implements IRateProvider {
+    private readonly url: string = config.BINANCE_ENDPOINT;
 
     public async getRate(): Promise<number> {
         try {
             const response = await axios.get(this.url);
-            return Number(response.data.data.amount);
+            return Number(response.data.price);
         } catch (error) {
             return super.getRate();
         }
