+++ src/components/rate/services/providers/coinmarket.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import config from '../../../config';
+import config from '../../../../config';
 
 import AbstractRateProvider, { IRateProvider } from './provider.interface';
