+++ src/components/subscription/__tests__/data/emails.test.json
