+++ src/components/subscription/__tests__/integration/subscription.integration.test.ts
@@ -1,25 +1,18 @@
 import express from 'express';
 import request from 'supertest';
 import bodyParser from 'body-parser';
-import router from '../../routes/router';
-import FileEmailRepository from '../../repository/email/file.repository';
 
-import config from '../../config';
+import router from '../../../../router';
+import FileEmailRepository from '../../repository/file.repository';
+import config from '../../../../config';
 
-describe('Integration test for API', () => {
+describe('Integration test for subscribe service', () => {
     const app = express();
     app.use(bodyParser.json());
     app.use(bodyParser.urlencoded({ extended: true }));
     app.use('/api', router);
 
-    test('GET api/rate - get BTC to UAH rate', async () => {
-        const { body, statusCode } = await request(app).get('/api/rate');
-
-        expect(body).toEqual(expect.any(Number));
-        expect(statusCode).toBe(200);
-    });
-
-    test('POST api/send - send emails to subscribers', async () => {
+    test('POST api/sendEmails - no emails to send', async () => {
         const { statusCode } = await request(app).post('/api/sendEmails');
 
         expect(statusCode).toBe(400);
@@ -33,7 +26,7 @@ describe('Integration test for API', () => {
         expect(statusCode).toBe(200);
     });
 
-    test('POST api/subscribe - subscribe to notification service', async () => {
+    test('POST api/subscribe - email already exists', async () => {
         const { statusCode } = await request(app).post('/api/subscribe').send({
             email: 'dhuchchin@gmail.com',
         });
