+++ tsconfig.json
@@ -5,7 +5,7 @@
     "esModuleInterop": true,                             
     "forceConsistentCasingInFileNames": true,
     "strict": true,                                      
-    "skipLibCheck": true
+    "skipLibCheck": true,
   },
   "include": ["src/**/*"],
   "exclude": ["node_modules"],
